requirements.txt
twisted
autobahn[twisted]
git+https://github.com/meejah/txtorcon.git
zope.interface
